+++ kernel/syscall.h
@@ -60,7 +60,7 @@ int syscall_##fn(P1 p1, P2 p2, P3 p3, P4 p4, P5 p5) \
 
 DECL_SYSCALL1(print, const char*)
 DECL_SYSCALL1(sleep, unsigned int)
-DECL_SYSCALL1(malloc, size_t)
+DECL_SYSCALL1(malloc, int)
 DECL_SYSCALL1(free, void*)
 DECL_SYSCALL1(exit, int)
 DECL_SYSCALL0(waitForKey)
